fix moving toPriorityUi from domain to presentation
