+++ app/src/main/java/com/amsterdam/cutetudee/presentation/component/chip/priority/PriorityUi.kt
@@ -5,6 +5,7 @@ import androidx.annotation.StringRes
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import com.amsterdam.cutetudee.R
+import com.amsterdam.cutetudee.domain.model.Task
 import com.amsterdam.cutetudee.presentation.theme.AppTheme
 import com.amsterdam.cutetudee.presentation.theme.ColorType
 
@@ -16,21 +17,29 @@ enum class PriorityUi(
     LOW(
         labelRes = R.string.priority_low,
         iconRes = R.drawable.trade_down_icon,
-        selectedContainerColorProvider = { AppTheme.color.greenAccent }
+        selectedContainerColorProvider = { AppTheme.color.greenAccent },
     ),
 
     MEDIUM(
         labelRes = R.string.priority_medium,
         iconRes = R.drawable.alert_icon,
-        selectedContainerColorProvider = { AppTheme.color.yellowAccent }
+        selectedContainerColorProvider = { AppTheme.color.yellowAccent },
     ),
 
     HIGH(
         labelRes = R.string.priority_high,
         iconRes = R.drawable.flag_icon,
-        selectedContainerColorProvider = { AppTheme.color.pinkAccent }
-    );
+        selectedContainerColorProvider = { AppTheme.color.pinkAccent },
+    ),
+    ;
 
     val selectedContainerColor: Color
         @Composable get() = selectedContainerColorProvider()
 }
+
+fun Task.Priority.toPriorityUi() =
+    when (this) {
+        Task.Priority.LOW -> PriorityUi.LOW
+        Task.Priority.MEDIUM -> PriorityUi.MEDIUM
+        Task.Priority.HIGH -> PriorityUi.HIGH
+    }
